- in progress

+++ packages/ketcher-core/src/application/editor/tools/Clear.ts
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  ***************************************************************************/
-import { CoreEditor, EditorHistory } from 'application/editor';
+import { CoreEditor, EditorHistory, SequenceMode } from 'application/editor';
 import { BaseTool } from 'application/editor/tools/Tool';
 
 class ClearTool implements BaseTool {
@@ -22,6 +22,10 @@ class ClearTool implements BaseTool {
   constructor(private editor: CoreEditor) {
     this.editor = editor;
     this.history = new EditorHistory(editor);
+
+    if (editor.mode instanceof SequenceMode) {
+      return;
+    }
 
     const modelChanges = this.editor.drawingEntitiesManager.deleteAllEntities();
     this.editor.renderersContainer.update(modelChanges);
